requirements.txt
tensorflow
selenium
python-telegram-bot